Move Origin, AttrChar & AttrField to attr
